refactor: renamed raster expression to rasterTransform component
